# app/backend/swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple crud with a little solid applied, using typescript and sequelize."
  version: "1.0.0"
  title: "Football API"
  contact:
    email: "murilorainho01@gmail.com"
host: "localhost:3001"
basePath: "/doc"
...
schemes:
- "http"
paths:
  /login:
    post:
      tags:
      - "login"
      summary: "Login to get access to the token"
      description: "Returns the user's object (without password) and the token"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "A object to find and validate user login"
        required: true
        schema:
          type: "object"
          properties:
            email:
              type: string
            password:
              type: string
      responses:
        "200":
          description: "Everything went right"
          schema:
            type: "object"
            properties:
              user:
                type: "object"
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  role:
                    type: string
                  email:
                    type: string
              token:
                type: string
        "401":
          description: "Invalid/missing input"
          schema:
            type: "object"
            properties:
              message:
                type: string
...
  /matchs:
    get:
      tags:
      - "matchs"
      summary: "Get all matches"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "inProgress"
        in: "query"
        description: "To filter matches in progress or not"
        required: false
        type: boolean
      responses:
        "200":
          description: "Everything went right"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                id:
                  type: number
                homeTeam:
                  type: number
                homeTeamGoals:
                  type: number
                awayTeam:
                  type: number
                awayTeamGoals:
                  type: number
                inProgress:
                  type: boolean
                homeClub:
                  type: "object"
                  properties:
                    id:
                      type: number
                    clubName:
                      type: string
                awayClub:
                  type: "object"
                  properties:
                    id:
                      type: number
                    clubName:
                      type: string
        "400":
          description: "Invalid 'inProgress' property in query"
          schema:
            type: "object"
            properties:
              message:
                type: string
    post:
      tags:
      - "matchs"
      summary: "Create a match"
      description: "If everything went right, returns the match's object"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        description: "Token to authorize"
        required: true
        type: string
      - in: "body"
        name: "body"
        description: "An object with all the attributes you need to create a match"
        required: true
        schema:
          type: "object"
          properties:
            homeTeam:
              type: number
            awayTeam:
              type: number
            homeTeamGoals:
              type: number
            awayTeamGoals:
              type: number
            inProgress:
              type: boolean
      responses:
        "201":
          description: "Everything went right"
          schema:
            type: "object"
            properties:
              homeTeam:
                type: number
              awayTeam:
                type: number
              homeTeamGoals:
                type: number
              awayTeamGoals:
                type: number
              inProgress:
                type: boolean
        "401":
          description: "Invalid/missing input"
          schema:
            type: "object"
            properties:
              message:
                type: string
...